remove misleading line
I found this line confusing.

Since the spec says `The Client SHOULD NOT include the Tus-Resumable header in the request and the Server MUST ignore the header`. I thought it'd make sense to rather remove that line from the example so it doesn't mislead anyone else. Please let me no if there's anything out of place. :)

*Question though:*
Does it mean the header `Tus-Resumable` should also be excluded from the response, since the server is expected to ignore it?