Spec for keeping Upload-Expires header current
For massive uploads, we don't want the expiration taking out a file that is being actively uploaded. So instead, add into the spec a way for the Upload-Expires header to be updated with each PATCH request, thus allowing us to prune abandoned uploads faster without causing problems to those still uploading.