Add optional Content-Location response header
... to inform the Client about the permanent location of completed
uploads.

This is motivated by the following use cases:

* TUS servers may not implement storage themselves, such that the
  `Location` returned for a creation (POST) request be only
  transitional.

* TUS servers may decide to change the storage location once the upload
  is complete, for example to

  * implement content-based naming (where the file name is some hash
    over the content)

  * name content based on type

  * or select paths by size

The `Content-Location` header is already defined by the HTTP RFCs, so,
stricly speaking, its use should be implicitly allowed and well defined.
Yet, for TUS, it makes sense to allow use of this response header for
additional HTTP verbs (e.g. `PATCH`) and it might help implementation
interoperability to point it out.