Creation: Move follow-up logic out of header description.
It was slightly confusing that a lot of the logic of how the length
eventually becomes known is written under a bullet point of

> "The request MUST include one of the following headers"

but other parts related to the same topic, such as
`creation-defer-length` in the server's `Tus-Extension` header
aren't.

With this change, the bullet point (b) is now focused on the header
and what it is allowed to contain, and the logic is moved to the prose.